+++ packages/scenes/src/querying/SceneQueryRunner.ts
@@ -38,6 +38,9 @@ import { filterAnnotations } from './layers/annotations/filterAnnotations';
 import { getEnrichedDataRequest } from './getEnrichedDataRequest';
 import { AdHocFilterSet } from '../variables/adhoc/AdHocFiltersSet';
 import { findActiveAdHocFilterSetByUid } from '../variables/adhoc/patchGetAdhocFilters';
+import { registerQueryWithController } from './registerQueryWithController';
+import { findActiveGroupByVariablesByUid } from '../variables/groupby/findActiveGroupByVariablesByUid';
+import { GroupByVariable } from '../variables/groupby/GroupByVariable';
 
 let counter = 100;
 
@@ -73,9 +76,13 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
   private _layerAnnotations?: DataFrame[];
   private _resultAnnotations?: DataFrame[];
 
-  // Closest filter set if found)
+  // Closest filter set if found
   private _adhocFilterSet?: AdHocFilterSet;
   private _adhocFilterSub?: Unsubscribable;
+
+  // Closest group by variable if found
+  private _groupBySource?: GroupByVariable;
+  private _groupBySourceSub?: Unsubscribable;
 
   public getResultsStream() {
     return this._results;
@@ -279,6 +286,10 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
       this._adhocFilterSub.unsubscribe();
     }
 
+    if (this._groupBySourceSub) {
+      this._groupBySourceSub.unsubscribe();
+    }
+
     this._variableValueRecorder.recordCurrentDependencyValuesForSceneObject(this);
   }
 
@@ -366,8 +377,8 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
       const datasource = this.state.datasource ?? findFirstDatasource(queries);
       const ds = await getDataSource(datasource, this._scopedVars);
 
-      if (!this._adhocFilterSet) {
-        this.findAndSubscribeToAdhocFilters(datasource?.uid);
+      if (!this._adhocFilterSet || !this._groupBySource) {
+        this.findAndSubscribeToAdhocSets(datasource?.uid);
       }
 
       const runRequest = getRunRequest();
@@ -375,26 +386,31 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
 
       writeSceneLog('SceneQueryRunner', 'Starting runRequest', this.state.key);
 
+      let stream = runRequest(ds, request);
+
       if (secondaryRequest) {
         // change subscribe callback below to pipe operator
-        this._querySub = forkJoin([runRequest(ds, request), runRequest(ds, secondaryRequest)])
-          .pipe(timeShiftQueryResponseOperator)
-          .subscribe(this.onDataReceived);
-      } else {
-        this._querySub = runRequest(ds, request).subscribe(this.onDataReceived);
+        stream = forkJoin([stream, runRequest(ds, secondaryRequest)]).pipe(timeShiftQueryResponseOperator);
       }
+
+      stream = stream.pipe(
+        registerQueryWithController({
+          type: 'data',
+          request,
+          origin: this,
+          cancel: () => this.cancelQuery(),
+        })
+      );
+
+      this._querySub = stream.subscribe(this.onDataReceived);
     } catch (err) {
       console.error('PanelQueryRunner Error', err);
 
-      const result = {
+      this.onDataReceived({
         ...emptyPanelData,
         ...this.state.data,
         state: LoadingState.Error,
         errors: [toDataQueryError(err)],
-      };
-
-      this.setState({
-        data: result,
       });
     }
   }
@@ -432,6 +448,11 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
     if (this._adhocFilterSet) {
       // @ts-ignore (Temporary ignore until we update @grafana/data)
       request.filters = this._adhocFilterSet.state.filters;
+    }
+
+    if (this._groupBySource) {
+      // @ts-ignore (Temporary ignore until we update @grafana/data)
+      request.groupByKeys = this._groupBySource.state.value;
     }
 
     request.targets = request.targets.map((query) => {
@@ -539,16 +560,26 @@ export class SceneQueryRunner extends SceneObjectBase<QueryRunnerState> implemen
   /**
    * Walk up scene graph and find the closest filterset with matching data source
    */
-  private findAndSubscribeToAdhocFilters(uid: string | undefined) {
-    const set = findActiveAdHocFilterSetByUid(uid);
+  private findAndSubscribeToAdhocSets(uid: string | undefined) {
+    const filters = findActiveAdHocFilterSetByUid(uid);
 
-    if (!set || set.state.applyMode !== 'same-datasource') {
-      return;
+    const groupByVariable = findActiveGroupByVariablesByUid(uid);
+
+    if (filters && filters.state.applyMode === 'same-datasource') {
+      if (!this._adhocFilterSet) {
+        // Subscribe to filter set state changes so that queries are re-issued when it changes
+        this._adhocFilterSet = filters;
+        this._adhocFilterSub = this._adhocFilterSet?.subscribeToState(() => this.runQueries());
+      }
     }
 
-    // Subscribe to filter set state changes so that queries are re-issued when it changes
-    this._adhocFilterSet = set;
-    this._adhocFilterSub = this._adhocFilterSet?.subscribeToState(() => this.runQueries());
+    if (groupByVariable && groupByVariable.state.applyMode === 'same-datasource') {
+      if (!this._groupBySource) {
+        // Subscribe to aggregations set state changes so that queries are re-issued when it changes
+        this._groupBySource = groupByVariable;
+        this._groupBySourceSub = this._groupBySource?.subscribeToState(() => this.runQueries());
+      }
+    }
   }
 }